Add MusicWithoutDelay.h to library.properties includes field
The includes field in library.properties is used to specify the #include directives which should be automatically added to the sketch after the user selects Sketch > Include Library > MusicWithoutDelay. Previously this action would only add the line:

#include <Tone.h>

to the sketch. Thus, MusicWithoutDelay.h must be one of the values specified in this field.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#libraryproperties-file-format